Rename upload errors file + modify verifyFileSize tests
